cope without mktemp
r48653

+++ common.sh
@@ -1,7 +1,11 @@
 require_tmpdir() {
   if [ -z "$OS_PROBER_TMP" ]; then
-    export OS_PROBER_TMP="$(mktemp -d /tmp/os-prober.XXXXXX)"
-    trap "rm -rf $OS_PROBER_TMP" EXIT HUP INT QUIT TERM
+    if type mktemp; then
+      export OS_PROBER_TMP="$(mktemp -d /tmp/os-prober.XXXXXX)"
+      trap "rm -rf $OS_PROBER_TMP" EXIT HUP INT QUIT TERM
+    else
+      export OS_PROBER_TMP=/tmp
+    fi
   fi
 }
